+++ docs/autonomy.md
@@ -75,7 +75,7 @@ These keys can be used for local deployments if you're using the default hardhat
 
 ```bash
 # fetch a service
-$ autonomy fetch valory/oracle_hardhat:0.1.0:bafybeie7kugbdzj5fvfcoomwg5h2r6alrvmh3aegpfm5itsc3r4qwhlkm4  --service
+$ autonomy fetch valory/oracle_hardhat:0.1.0:bafybeibi6n2azy43b4g7ov4aujprj3pufrsiupgqjed7eer2ctczcuqusu  --service
 $ cd oracle_hardhat
 # create a docker deployment
 $ autonomy deploy build keys.json
