+++ docs/price_oracle_intro.md
@@ -41,7 +41,7 @@ you have followed the [setup instructions](guides/quick_start.md#setup). As a re
 
 2. Use the CLI to download the `valory/oracle_hardhat` service.
     ```bash
-    autonomy fetch valory/oracle_hardhat:0.1.0:bafybeie7kugbdzj5fvfcoomwg5h2r6alrvmh3aegpfm5itsc3r4qwhlkm4 --remote --service
+    autonomy fetch valory/oracle_hardhat:0.1.0:bafybeibi6n2azy43b4g7ov4aujprj3pufrsiupgqjed7eer2ctczcuqusu --remote --service
     cd oracle_hardhat
     ```
