+++ packages/packages.json
@@ -22,7 +22,7 @@
     "contract/valory/uniswap_v2_erc20/0.1.0": "bafybeibkiwytuhhvxbisoxcybrx4lfstk6bvtriltles7fxxpy37yq45ja",
     "contract/valory/uniswap_v2_router_02/0.1.0": "bafybeidabycj7yw6u2bt64db23lvb5zowmcgu6zjkp7xjigqle3y3scgfi",
     "skill/valory/apy_estimation_abci/0.1.0": "bafybeiex4atuk2rcy3e2aj5wmmmg6tnfsvvjgvotqo3c7xu7ayi6lwz2xq",
-    "skill/valory/price_estimation_abci/0.1.0": "bafybeigthpgeb6jgxz3agymonbrgxn3rqjzdunuklj5mmoz3v4437imdfm",
+    "skill/valory/price_estimation_abci/0.1.0": "bafybeidooi74zt4ta2pl7l4pktzdm7os5wcivblyohw3rkoylt6z6trnti",
     "skill/valory/registration_abci/0.1.0": "bafybeiajvfkffbdmggj3it7e4r6733lylxhbnb2a6tx3tmlgpy2fpxkuze",
     "skill/valory/reset_pause_abci/0.1.0": "bafybeieo5m37rwoel7e6u3lrp2cei7hdyhklggsc6776vej6bp6gzua7ei",
     "skill/valory/safe_deployment_abci/0.1.0": "bafybeifkxk6jkixnk4fymsw76zw2z2ofshmomqbxkp2kcwpur6dn7jb44i",
@@ -31,7 +31,7 @@
     "skill/valory/counter_client/0.1.0": "bafybeigbq3tkjojxognv3auyhablqd6lybeahc6xd6rnpi4dlw4jryedwq",
     "skill/valory/hello_world_abci/0.1.0": "bafybeiamhn2ri7ocdummqhwgd75lqeev3ruviqdfuujae4adtksmleetmy",
     "skill/valory/liquidity_rebalancing_abci/0.1.0": "bafybeidcwmmzah2dijdjjmt3fqgxms2iinshloumife5hxoohtza4nrcum",
-    "skill/valory/oracle_abci/0.1.0": "bafybeidggop5kc4b65qxtvdyzsmohpv6flqxr2ybxfglngbmcjve76ggym",
+    "skill/valory/oracle_abci/0.1.0": "bafybeiaq7x5a3mukydzqrgbf4iolux3ken24pf46nj44ak4paii2ls6ls4",
     "skill/valory/register_reset_abci/0.1.0": "bafybeiawo72hspab7nyt7gcjs7b44qksem26mhmkfriuov7rhebtkpnxre",
     "skill/valory/simple_abci/0.1.0": "bafybeifw3msuw67ird6hbatosh2yhql5n5k44dlwnyjwlek3g6ergpwpza",
     "skill/valory/test_abci/0.1.0": "bafybeigrqbjnjqhbcifoostknq3qxyox3rg2v7z76m22iifpj45pl6l66a",
@@ -40,7 +40,7 @@
     "agent/valory/counter/0.1.0": "bafybeidqj3gwaageeqwi62ymxfvhkh4lfjmjmra2b3ayjlwdg3bltyye3m",
     "agent/valory/counter_client/0.1.0": "bafybeiaerrvu26a2cxzc4baatqg4zha7z5m3skxzoakfx43eyweveizhgm",
     "agent/valory/hello_world/0.1.0": "bafybeicqhwxerp5codhdpiobfv3aaquig2qbci442yvvf65ziuenaad3wq",
-    "agent/valory/oracle/0.1.0": "bafybeihcomi2hqq4l3nkrjtzkv3npni6zyaqhfs2bwkvwh3azrapvgalom",
+    "agent/valory/oracle/0.1.0": "bafybeihqkyrabznuwsmki3ou43f43tph2h3rjo5hk5mbn6memj5lxi3kru",
     "agent/valory/register_reset/0.1.0": "bafybeiaw2dc462bvy5ycrcdd22juhzrzf3zzi4u6ls5e2qa4ml4o3skxcm",
     "agent/valory/registration_start_up/0.1.0": "bafybeicylq7okzw4lw3be4ixgqr6pkrnlfi63enawfwznqxconrxgrnjqm",
     "agent/valory/simple_abci/0.1.0": "bafybeidfv553nxcyqgg7s4kosxoseba5ru7ygm3dsy7fuierdstesmjraq",
@@ -49,11 +49,11 @@
     "service/valory/apy_estimation_hardhat/0.1.0": "bafybeiefrtifcffxbxasqtzsw6zhq3t5tf4ubt2hvpzalc7hczr2iwcg34",
     "service/valory/counter/0.1.0": "bafybeiajvvk6qcr7dorjrgjti465d6p5auhbsgmpjf4pwtmgfgwjae7vom",
     "service/valory/hello_world/0.1.0": "bafybeihrng5d73bsctkjilgc2lxkn4ditsbf3c7vqmt6wpi45zglq42i2e",
-    "service/valory/oracle_goerli/0.1.0": "bafybeihp2huisurreroh2cmz5syf3hzspoba7d34t7pajqpqzudnefgply",
-    "service/valory/oracle_hardhat/0.1.0": "bafybeie7kugbdzj5fvfcoomwg5h2r6alrvmh3aegpfm5itsc3r4qwhlkm4",
-    "service/valory/oracle_polygon/0.1.0": "bafybeidhpj2ouetebmqrhybhg3pxmgylcsidijty6fcpeyng34gwi4a24i",
-    "service/valory/oracle_ropsten/0.1.0": "bafybeiaeedzj3cg67xtnncnpqx2fs3mz4gov6rk7yujfzoveibaxsef6ui",
-    "service/valory/oracle_ropsten_cluster/0.1.0": "bafybeia6svlydxp4soluuz6tey372qiw6a47oujdryoxmh3xtfnxta22s4",
+    "service/valory/oracle_goerli/0.1.0": "bafybeif3bymikyz6r3rhshoo6tsz4mtwcbafiycm5lepanh6uba4iv65rm",
+    "service/valory/oracle_hardhat/0.1.0": "bafybeibi6n2azy43b4g7ov4aujprj3pufrsiupgqjed7eer2ctczcuqusu",
+    "service/valory/oracle_polygon/0.1.0": "bafybeiahxpn6wacovcvoxhf2fmvmz7c2cnadtboqjja3vcsswpkimvrueu",
+    "service/valory/oracle_ropsten/0.1.0": "bafybeibeviksmwtfetd4qd4zfniggacchz2vq4x24se2lxwvfpsigaa2wm",
+    "service/valory/oracle_ropsten_cluster/0.1.0": "bafybeidn5dohcvrprsgg6m42fggbdw5vpesrrd566foeq2ihakaiqle72y",
     "service/valory/register_reset/0.1.0": "bafybeidkzsisq62mmd4xdfv4joxpnyvyp3rihd7mpiseddvyb6wwoul7ei",
     "service/valory/simple_abci/0.1.0": "bafybeibcplvpk56a47vgblnhpji435ennnsn64ohdebewwefra7gnywxpa",
     "skill/valory/liquidity_provision_abci/0.1.0": "bafybeigjxgwzovmg4p3l2ewn43ai62ybc4lwjj3wsqfuys4yi2gbscefia"
